Refined approach to LinkList Challenge
Create the structure as follows: head, prev and tail all point to nodes out of range of the list (a negative number for example) current points to tail. This the new "empty". Avoids having to check for empty, 1 node, or 2 nodes in the insert methods. Currently at 10 lines of code down from 27!